Convert Search WorkItem to JSON API
Related to #526
Related to #515
