+++ design/jsonapi_helpers.go
@@ -78,8 +78,6 @@ var genericData = a.Type("GenericData", func() {
 	a.Attribute("id", d.String, "UUID of the object", func() {
 		a.Example("6c5610be-30b2-4880-9fec-81e4f8e4fd76")
 	})
-	a.Required("type")
-	a.Required("id")
 })
 
 // genericLinks defines generic relations links
