+++ design/workitems.go
@@ -42,8 +42,7 @@ var assigneeData = a.Type("AssigneeData", func() {
 	a.Attribute("id", d.String, "UUID of the identity", func() {
 		a.Example("6c5610be-30b2-4880-9fec-81e4f8e4fd76")
 	})
-	a.Required("type")
-	a.Required("id")
+	a.Required("type", "id")
 })
 
 // relationBaseType is top level block for WorkItemType relationship
